Fix pytest
