+++ src/imitation/trainer.py
@@ -100,9 +100,9 @@ class Trainer:
         self.env_train = self.env_test = self.env
     else:
         self.env_train = reward_wrapper.RewardVecEnvWrapper(
-            self.env, self.discrim.reward_train)
+            self.env, self.reward_train)
         self.env_test = reward_wrapper.RewardVecEnvWrapper(
-            self.env, self.discrim.reward_test)
+            self.env, self.reward_test)
 
     if gen_replay_buffer_capacity is None:
         gen_replay_buffer_capacity = 20 * self._n_disc_samples_per_buffer
@@ -285,6 +285,61 @@ class Trainer:
         self.discrim.log_policy_act_prob_ph: log_act_prob,
     }
     return fd
+
+  def reward_train(
+    self,
+    old_obs: np.ndarray,
+    act: np.ndarray,
+    new_obs: np.ndarray,
+  ) -> np.ndarray:
+    """Vectorized reward for training an imitation learning algorithm.
+
+    Args:
+        old_obs (array): The observation input. Its shape is
+            `((None,) + observation_space.shape)`.
+        act (array): The action input. Its shape is
+            `((None,) + action_space.shape)`. The None dimension is
+            expected to be the same as None dimension from `obs_input`.
+        new_obs (array): The observation input. Its shape is
+            `((None,) + observation_space.shape)`.
+    """
+    old_obs = np.atleast_1d(old_obs)
+    act = np.atleast_1d(act)
+    new_obs = np.atleast_1d(new_obs)
+
+    n_gen = len(old_obs)
+    assert len(act) == n_gen
+    assert len(new_obs) == n_gen
+
+    # Calculate generator-policy log probabilities.
+    log_act_prob = self._gen_policy.action_probability(old_obs, actions=act,
+                                                       logp=True)
+    assert len(log_act_prob) == n_gen
+    log_act_prob = log_act_prob.reshape((n_gen,))
+
+    fd = {
+        self.discrim.old_obs_ph: old_obs,
+        self.discrim.act_ph: act,
+        self.discrim.new_obs_ph: new_obs,
+        self.discrim.labels_ph: np.ones(n_gen),
+        self.discrim.log_policy_act_prob_ph: log_act_prob,
+    }
+    rew = self._sess.run(self._discrim.policy_train_reward, feed_dict=fd)
+    return rew.flatten()
+
+  def reward_test(self, *args, **kwargs) -> np.ndarray:
+    """Vectorized reward for training an expert during transfer learning.
+
+    Args:
+        old_obs (array): The observation input. Its shape is
+            `((None,) + observation_space.shape)`.
+        act (array): The action input. Its shape is
+            `((None,) + action_space.shape)`. The None dimension is
+            expected to be the same as None dimension from `obs_input`.
+        new_obs (array): The observation input. Its shape is
+            `((None,) + observation_space.shape)`.
+    """
+    return self.discrim.reward_test(*args, **kwargs)
 
 
 def _n_steps_if_not_none(n_steps):
